+++ src/gallery-flickr-carousel/HISTORY.md
@@ -1,3 +1,10 @@
+YUI Flickr Carousel
+===================
+
+### 0.2.0 - January 2 2013
+
+*   Converted module to use yogi
+
 ### 0.1.1 - January 2 2013
 
 *   Switched from using Get to JSONP